requirements.txt
streamlit>=1.32.0
sentence-transformers>=2.7.0
faiss-cpu>=1.8.0
groq>=0.5.0
requests>=2.31.0
numpy>=1.26.0
audio-recorder-streamlit>=0.0.8
python-dotenv>=1.0.0
